lib/, src/: Remove useless casts in fgets(3)
This patch can be replicated with the following semantic patch:

	$ cat fgets_cast.sp
	@@
	expression a, b, c;
	@@

	- fgets(a, (int) (b), c)
	+ fgets(a, b, c)

	@@
	expression a, b, c;
	@@

	- fgets(a, (int) b, c)
	+ fgets(a, b, c)

	@@
	expression a, b, c;
	@@

	- fgetsx(a, (int) (b), c)
	+ fgetsx(a, b, c)

	@@
	expression a, b, c;
	@@

	- fgetsx(a, (int) b, c)
	+ fgetsx(a, b, c)

	@@
	expression a, b, c, p;
	@@

	- p->fgets(a, (int) (b), c)
	+ p->fgets(a, b, c)

	@@
	expression a, b, c, p;
	@@

	- p->fgets(a, (int) b, c)
	+ p->fgets(a, b, c)

which is applied as:

	$ find lib* src/ -type f \
	| xargs spatch --sp-file ~/tmp/spatch/fgets_cast.sp --in-place;

+++ lib/gshadow.c
@@ -170,9 +170,7 @@ void endsgent (void)
 		buflen *= 2;
 
 		len = strlen (buf);
-		if (fgetsx (&buf[len],
-			    (int) (buflen - len),
-			    fp) != &buf[len]) {
+		if (fgetsx(&buf[len], buflen - len, fp) != &buf[len]) {
 			return NULL;
 		}
 	}
